Fix integration tests pylint
