+++ tests/Integration/AccountingSystem/Test_DataStoreClient.py
@@ -19,40 +19,10 @@ from DIRAC import gLogger
 from DIRAC.AccountingSystem.Client.DataStoreClient import gDataStoreClient
 from DIRAC.AccountingSystem.Client.Types.DataOperation import DataOperation
 from DIRAC.AccountingSystem.Client.Types.StorageOccupancy import StorageOccupancy
+from DIRAC.tests.Utilities.Accounting import createDataOperationAccountingRecord
+from DIRAC.tests.Utilities.Accounting import createStorageOccupancyAccountingRecord
 
 gLogger.setLevel('DEBUG')
-
-
-def createDataOperationAccountingRecord():
-  accountingDict = {}
-  accountingDict['OperationType'] = 'putAndRegister'
-  accountingDict['User'] = 'system'
-  accountingDict['Protocol'] = 'DataManager'
-  accountingDict['RegistrationTime'] = 0.0
-  accountingDict['RegistrationOK'] = 0
-  accountingDict['RegistrationTotal'] = 0
-  accountingDict['Destination'] = 'se'
-  accountingDict['TransferTotal'] = 1
-  accountingDict['TransferOK'] = 1
-  accountingDict['TransferSize'] = 1
-  accountingDict['TransferTime'] = 0.0
-  accountingDict['FinalStatus'] = 'Successful'
-  accountingDict['Source'] = 'testSite'
-  oDataOperation = DataOperation()
-  oDataOperation.setValuesFromDict(accountingDict)
-  return oDataOperation
-
-
-def createStorageOccupancyAccountingRecord():
-  accountingDict = {}
-  accountingDict['Site'] = 'LCG.PIPPO.it'
-  accountingDict['Endpoint'] = 'somewhere.in.topolinea.it'
-  accountingDict['StorageElement'] = 'PIPPO-SE'
-  accountingDict['SpaceType'] = 'Total'
-  accountingDict['Space'] = 123456
-  oSO = StorageOccupancy()
-  oSO.setValuesFromDict(accountingDict)
-  return oSO
 
 
 def test_addAndRemoveDataperation():
