+++ tests/Integration/Monitoring/Test_Plotter.py
@@ -10,7 +10,7 @@ from __future__ import print_function
 
 import os
 
-from DIRAC.tests.Integration.AccountingSystem.Test_Plots import compare
+from DIRAC.tests.Utilities.plots import compare
 
 # sut
 from DIRAC.MonitoringSystem.private.Plotters.ComponentMonitoringPlotter import ComponentMonitoringPlotter
